tests/test_emissor/test_interaction/lets-chat-requirements.txt
jsonpickle
numpy
opencv-python
Pillow
python-on-whales
requests
spacy
tqdm
